thumbnails/requirements.txt
kafka-python
requests
boto3